refactor: move types from template to runtime
