+++ src/runtime/index.ts
@@ -2,3 +2,4 @@ export * from './functions/table-dispatcher';
 export * from './functions/model-change-handler';
 export * from './assert';
 export * from './errors';
+export * from './types';
