+++ src/runtime/types.ts
@@ -1,0 +1,31 @@
+import type {
+  ConsumedCapacity,
+  ItemCollectionMetrics,
+} from '@aws-sdk/client-dynamodb';
+import type {NativeAttributeValue} from '@aws-sdk/util-dynamodb';
+
+export interface QueryOptions {
+  limit?: number;
+  nextToken?: Record<string, NativeAttributeValue>;
+  /**
+   * All operators supported by DynamoDB are except `between`. `between` is
+   * not supported because it requires two values and that makes the codegen
+   * quite a bit more tedious. If it's needed, please open a ticket and we can
+   * look into adding it.
+   */
+  operator?: 'begins_with' | '=' | '<' | '<=' | '>' | '>=';
+  reverse?: boolean;
+}
+
+export interface ResultType<T> {
+  capacity: ConsumedCapacity;
+  item: T;
+  metrics: ItemCollectionMetrics | undefined;
+}
+
+export interface MultiResultType<T> {
+  capacity: ConsumedCapacity;
+  hasNextPage: boolean;
+  items: T[];
+  nextToken: Record<string, NativeAttributeValue> | undefined;
+}
